Изменил PropertyHandler и BookingHandler чтобы они возвращали id а не весь ентити в контроллер

+++ Presenter/DTOs/PropertyDTO/CreatePropertyRequest.cs
@@ -104,7 +104,7 @@ namespace Presenter.DTOs.PropertyDTO
         /// </summary>
         [Required(ErrorMessage = "Площадь обязательна")]
         [Range(1, double.MaxValue, ErrorMessage = "Площадь должна быть больше 0")]
-        public double Area { get; init; }
+        public decimal Area { get; init; }
 
         /// <summary>
         /// Наличие парковки
